Extracted CharacterPresetEditor from components.js
