+++ src/js/shot-generator-new/components/CharacterPresetEditor/index.js
@@ -1,0 +1,153 @@
+const React = require('react')
+const { connect } = require('react-redux')
+const {
+    updateObject,
+
+    createCharacterPreset,
+
+    undoGroupStart,
+    undoGroupEnd,
+  } = require('../../../shared/reducers/shot-generator')
+const presetsStorage = require('../../../shared/store/presetsStorage')
+const h = require('../../../utils/h')
+
+const preventDefault = (fn, ...args) => e => {
+    e.preventDefault()
+    fn(e, ...args)
+  }
+const saveCharacterPresets = state => presetsStorage.saveCharacterPresets({ characters: state.presets.characters })
+const shortId = id => id.toString().substr(0, 7).toLowerCase()
+
+const CharacterPresetsEditor = connect(
+  state => ({
+    characterPresets: state.presets.characters,
+    models: state.models
+  }),
+  {
+    updateObject,
+    selectCharacterPreset: (sceneObject, characterPresetId, preset) => (dispatch, getState) => {
+      dispatch(updateObject(sceneObject.id, {
+        // set characterPresetId
+        characterPresetId,
+
+        // apply preset values to character model
+        height: preset.state.height,
+        //height: state.models[preset.state.model].baseHeight,
+        model: preset.state.model,
+        // gender: 'female',
+        // age: 'adult'
+
+        headScale: preset.state.headScale,
+        tintColor: preset.state.tintColor,
+
+        morphTargets: {
+          mesomorphic: preset.state.morphTargets.mesomorphic,
+          ectomorphic: preset.state.morphTargets.ectomorphic,
+          endomorphic: preset.state.morphTargets.endomorphic
+        },
+
+        name: sceneObject.name || preset.name
+      }))
+    },
+    createCharacterPreset: ({ id, name, sceneObject }) => (dispatch, getState) => {
+      // add the character data to a named preset
+      let preset = {
+        id,
+        name,
+        state: {
+          height: sceneObject.height,
+          //height: sceneObject.model.originalHeight,
+
+          model: sceneObject.model,
+          // gender: 'female',
+          // age: 'adult'
+
+          headScale: sceneObject.headScale,
+          tintColor: sceneObject.tintColor,
+
+          morphTargets: {
+            mesomorphic: sceneObject.morphTargets.mesomorphic,
+            ectomorphic: sceneObject.morphTargets.ectomorphic,
+            endomorphic: sceneObject.morphTargets.endomorphic
+          }
+        }
+      }
+
+      // start the undo-able operation
+      dispatch(undoGroupStart())
+
+      // create the preset
+      dispatch(createCharacterPreset(preset))
+
+      // save the presets file
+      saveCharacterPresets(getState())
+
+      // update this object to use the preset
+      dispatch(updateObject(sceneObject.id, {
+        // set the preset id
+        characterPresetId: id,
+        // use the preset’s name (if none assigned)
+        name: sceneObject.name || preset.name
+      }))
+
+      // end the undo-able operation
+      dispatch(undoGroupEnd())
+    }
+  }
+)(
+  // TODO could optimize by only passing sceneObject properties we actually care about
+  React.memo(({ sceneObject, characterPresets, selectCharacterPreset, createCharacterPreset }) => {
+      console.log("render")
+    const onCreateCharacterPresetClick = event => {
+      // show a prompt to get the desired preset name
+      let id = THREE.Math.generateUUID()
+      prompt({
+        title: 'Preset Name',
+        label: 'Select a Preset Name',
+        value: `Character ${shortId(id)}`
+      }, require('electron').remote.getCurrentWindow()).then(name => {
+        if (name != null && name != '' && name != ' ') {
+          createCharacterPreset({
+            id,
+            name,
+            sceneObject
+          })
+        }
+      }).catch(err => {
+        console.error(err)
+      })
+    }
+
+    const onSelectCharacterPreset = event => {
+      let characterPresetId = event.target.value
+      let preset = characterPresets[characterPresetId]
+      selectCharacterPreset(sceneObject, characterPresetId, preset)
+    }
+
+    return h(
+      ['div.row', { style: { margin: '9px 0 6px 0', paddingRight: 0 } }, [
+        ['div', { style: { width: 50, display: 'flex', alignSelf: 'center' } }, 'preset'],
+        [
+          'select', {
+            required: true,
+            value: sceneObject.characterPresetId || '',
+            onChange: preventDefault(onSelectCharacterPreset),
+            style: {
+              flex: 1,
+              marginBottom: 0,
+              maxWidth: 192
+            }
+          }, [
+              ['option', { value: '', disabled: true }, '---'],
+              Object.values(characterPresets).map(preset =>
+                ['option', { value: preset.id }, preset.name]
+              )
+            ]
+          ]
+        ],
+        ['a.button_add[href=#]', { style: { marginLeft: 6 }, onClick: preventDefault(onCreateCharacterPresetClick) }, '+']
+      ]
+    )
+  })
+)
+module.exports = CharacterPresetsEditor
